Nicolas Regnault: - add AbstractScalarSumOperation class to handle all operation whose results is the sum of indepdendant scalar contributions. - add QHEParticleWaveFunctionOperation to parallelize evaluation of QHE wave function in real space
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@135 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ src/Matrix/ComplexMatrix.h
@@ -269,13 +269,13 @@ class ComplexMatrix : protected Matrix
   // changeBit = array indicating which bit is changed at the i-th iteration of the Gray code
   // changeBitSign = array with -1 if the changed bit is from 1 to 0, +1 either
   // return value = permanent associated to the matrix
-  Complex FastPermanent(int* changeBit, double* changeBitSign);
+  Complex FastPermanent(int* changeBit, int* changeBitSign);
 
   // evaluate precalculation array  neede for the fast permanent calculation
   //
   // changeBit = reference on the array indicating which bit is changed at the i-th iteration of the Gray code
   // changeBitSign = reference on array with -1 if the changed bit is from 1 to 0, +1 either
-  void EvaluateFastPermanentPrecalculationArray(int*& changeBit, double*& changeBitSign);
+  void EvaluateFastPermanentPrecalculationArray(int*& changeBit, int*& changeBitSign);
 
   // Output Stream overload
   //
